+++ CHANGELOG.md
@@ -91,6 +91,8 @@
   - Style cleanup in code and tests.
 - Changes to SqlScript
   - Style cleanup in tests.
+  - Updated examples.
+  - Added integration tests.
 - Changes to SqlServerConfiguration
   - Fixed minor typos in comment-based help.
   - Now the verbose message say what option is changing and to what value